Rename to test-pyramid-app
The repo has been renamed to test-pyramid-app. Update lots of references
to the old app name in the repo contents.
