+++ .cookiecutter/cookiecutter.json
@@ -4,8 +4,8 @@
   "ignore": [],
   "extra_context": {
     "name": "Pyramid App Cookiecutter Test",
-    "package_name": "pyramid_app_cookiecutter_test",
-    "slug": "pyramid-app-cookiecutter-test",
+    "package_name": "test_pyramid_app",
+    "slug": "test-pyramid-app",
     "short_description": "A test app for our pyramid-app cookiecutter.",
     "python_version": "3.10.4",
     "port": "9800",
@@ -15,10 +15,10 @@
     "dependabot_pip_interval": "monthly",
     "dependabot_npm_interval": "monthly",
     "dependabot_docker_interval": "monthly",
-    "__github_url": "https://github.com/hypothesis/pyramid-app-cookiecutter-test",
+    "__github_url": "https://github.com/hypothesis/test-pyramid-app",
     "docker": "yes",
     "__docker_namespace": "hypothesis",
-    "__docker_network": "pyramid_app_cookiecutter_test_default",
+    "__docker_network": "test_pyramid_app_default",
     "postgres": "yes",
     "frontend": "no",
     "__postgres_port": "5438",
