+++ .github/workflows/ci.yml
@@ -88,7 +88,7 @@ jobs:
           restore-keys: |
             tests-${{ runner.os }}-tox-
       - name: Create test database
-        run: psql -U postgres -h localhost -p 5438 -c 'CREATE DATABASE pyramid_app_cookiecutter_test_test'
+        run: psql -U postgres -h localhost -p 5438 -c 'CREATE DATABASE test_pyramid_app_test'
       - run: python -m pip install 'tox<4'
       - run: tox -e tests
       - name: Upload coverage file
@@ -145,6 +145,6 @@ jobs:
           restore-keys: |
             functests-${{ runner.os }}-tox-
       - name: Create test database
-        run: psql -U postgres -h localhost -p 5438 -c 'CREATE DATABASE pyramid_app_cookiecutter_test_test'
+        run: psql -U postgres -h localhost -p 5438 -c 'CREATE DATABASE test_pyramid_app_test'
       - run: python -m pip install 'tox<4'
       - run: tox -e functests
