+++ README.md
@@ -1,5 +1,5 @@
 <a><img src="https://img.shields.io/badge/python-3.10-success"></a>
-<a href="https://github.com/hypothesis/pyramid-app-cookiecutter-test/blob/main/LICENSE"><img src="https://img.shields.io/badge/license-BSD--2--Clause-success"></a>
+<a href="https://github.com/hypothesis/test-pyramid-app/blob/main/LICENSE"><img src="https://img.shields.io/badge/license-BSD--2--Clause-success"></a>
 <a href="https://github.com/hypothesis/cookiecutters/tree/main/pyramid-app"><img src="https://img.shields.io/badge/cookiecutter-pyramid--app-success"></a>
 <a href="https://black.readthedocs.io/en/stable/"><img src="https://img.shields.io/badge/code%20style-black-000000"></a>
 
@@ -28,8 +28,8 @@ First you'll need to install:
 Then to set up your development environment:
 
 ```terminal
-git clone https://github.com/hypothesis/pyramid-app-cookiecutter-test.git
-cd pyramid-app-cookiecutter-test
+git clone https://github.com/hypothesis/test-pyramid-app.git
+cd test-pyramid-app
 make services
 make help
 ```
