+++ conf/development.ini
@@ -1,12 +1,12 @@
 [app:main]
-use = call:pyramid_app_cookiecutter_test.app:create_app
+use = call:test_pyramid_app.app:create_app
 debug = true
 
 [pshell]
-setup = pyramid_app_cookiecutter_test.pshell.setup
+setup = test_pyramid_app.pshell.setup
 
 [loggers]
-keys = root, pyramid_app_cookiecutter_test
+keys = root, test_pyramid_app
 
 [handlers]
 keys = console
@@ -18,10 +18,10 @@ keys = generic
 level = INFO
 handlers = console
 
-[logger_pyramid_app_cookiecutter_test]
+[logger_test_pyramid_app]
 level = DEBUG
 handlers =
-qualname = pyramid_app_cookiecutter_test
+qualname = test_pyramid_app
 
 [handler_console]
 class = StreamHandler
