+++ conf/production.ini
@@ -1,16 +1,16 @@
 [pipeline:main]
 pipeline:
   proxy-prefix
-  pyramid_app_cookiecutter_test
+  test_pyramid_app
 
-[app:pyramid_app_cookiecutter_test]
-use = call:pyramid_app_cookiecutter_test.app:create_app
+[app:test_pyramid_app]
+use = call:test_pyramid_app.app:create_app
 
 [filter:proxy-prefix]
 use: egg:PasteDeploy#prefix
 
 [loggers]
-keys = root, pyramid_app_cookiecutter_test, alembic
+keys = root, test_pyramid_app, alembic
 
 [handlers]
 keys = console
@@ -22,10 +22,10 @@ keys = generic
 level = INFO
 handlers = console
 
-[logger_pyramid_app_cookiecutter_test]
+[logger_test_pyramid_app]
 level = DEBUG
 handlers =
-qualname = pyramid_app_cookiecutter_test
+qualname = test_pyramid_app
 
 [logger_alembic]
 level = INFO
