+++ conf/supervisord-dev.conf
@@ -25,4 +25,4 @@ supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
 
 [supervisorctl]
 serverurl = unix://.supervisor.sock
-prompt = pyramid-app-cookiecutter-test
+prompt = test-pyramid-app
