+++ tox.ini
@@ -13,8 +13,8 @@ setenv =
     PYTHONUNBUFFERED = 1
     OBJC_DISABLE_INITIALIZE_FORK_SAFETY = YES
     dev: DATABASE_URL = {env:DATABASE_URL:postgresql://postgres@localhost:5438/postgres}
-    tests,functests: TEST_DATABASE_URL = {env:TEST_DATABASE_URL:postgresql://postgres@localhost:5438/pyramid_app_cookiecutter_test_test}
-    # Make `import pyramid_app_cookiecutter_test` work in `make shell`.
+    tests,functests: TEST_DATABASE_URL = {env:TEST_DATABASE_URL:postgresql://postgres@localhost:5438/test_pyramid_app_test}
+    # Make `import test_pyramid_app` work in `make shell`.
     dev: PYTHONPATH = .
 passenv =
     HOME
@@ -32,17 +32,17 @@ allowlist_externals =
 commands_pre =
     pip-sync requirements/{env:TOX_ENV_NAME}.txt --pip-args '--disable-pip-version-check'
 commands =
-    tests: sh bin/create-db pyramid_app_cookiecutter_test_test
-    functests: sh bin/create-db pyramid_app_cookiecutter_test_functests
+    tests: sh bin/create-db test_pyramid_app_test
+    functests: sh bin/create-db test_pyramid_app_functests
     dev: {posargs:supervisord -c conf/supervisord-dev.conf}
-    format: black pyramid_app_cookiecutter_test tests bin
-    format: isort --atomic pyramid_app_cookiecutter_test tests bin
-    checkformatting: black --check pyramid_app_cookiecutter_test tests bin
-    checkformatting: isort --quiet --check-only pyramid_app_cookiecutter_test tests bin
-    lint: pylint pyramid_app_cookiecutter_test bin
+    format: black test_pyramid_app tests bin
+    format: isort --atomic test_pyramid_app tests bin
+    checkformatting: black --check test_pyramid_app tests bin
+    checkformatting: isort --quiet --check-only test_pyramid_app tests bin
+    lint: pylint test_pyramid_app bin
     lint: pylint --rcfile=tests/pyproject.toml tests
-    lint: pydocstyle pyramid_app_cookiecutter_test tests bin
-    lint: pycodestyle pyramid_app_cookiecutter_test tests bin
+    lint: pydocstyle test_pyramid_app tests bin
+    lint: pycodestyle test_pyramid_app tests bin
     tests: coverage run -m pytest --failed-first --new-first --no-header --quiet {posargs:tests/unit/}
     functests: pytest --failed-first --new-first --no-header --quiet {posargs:tests/functional/}
     coverage: -coverage combine
